pg_metrics_consumer/requirements.txt
confluent-kafka[avro]==2.3.0
psycopg2-binary==2.9.9
